Fix clippy warnings: wildcards, map_err, numeric fallback, as_conversions
- Replace wildcard enum matches with explicit XBinary::NegInf | XBinary::PosInf
- Add From<IntervalError> for ComputableError, replacing map_err(|_|...)
- Add type suffixes to numeric literals (default_numeric_fallback)
- Change MAX_COMPUTATION_BITS from u64 to usize (reflects memory limits)
- Change pi_bounds_at_precision and related fns from u64 to usize
- Add bits_as_usize() helper centralizing the u64-from-.bits()-to-usize
  conversion with a MAX_COMPUTATION_BITS sanity check
- Make Computable.node private, add from_node() constructor
- Use try_from/saturating arithmetic to eliminate unsafe as casts
- Simplify assert_sane_computation_size! macro to take usize directly

+++ src/binary_utils/bisection.rs
@@ -246,7 +246,7 @@ pub fn normalize_bounds(
         width_ubinary.exponent().clone()
     } else {
         // Add 1 extra to account for rounding when flooring the lower bound
-        width_ubinary.exponent() + BigInt::from(width_bits as i64) + BigInt::one()
+        width_ubinary.exponent() + BigInt::from(width_bits) + BigInt::one()
     };
 
     // Floor the lower bound to this exponent
